refactor: split attribution set results
