+++ quality/baseline_report.md
@@ -13,11 +13,11 @@ completion claim.
 ## Generation Identity
 
 - Git branch: `feat/enterprise-risk-refactor-continuation`
-- Git commit: `5141005749ae60b155ada8e1d285bf2a966d10c9`
+- Git commit: `b707e7f4ed27caa969a86db7c9a818d2d20e748a`
 
 ## Current Code Size
 
-- Python source files under `src/`: 220
+- Python source files under `src/`: 221
 - Python test files under `tests/`: 103
 - Python packages under `src/`: 9
 - API entry point route/middleware/handler decorators in `src/app/main.py`: 0
@@ -26,7 +26,6 @@ completion claim.
 
 | Path | Lines | Bytes |
 | --- | --- | --- |
-| src/app/services/attribution_decomposition.py | 279 | 9057 |
 | src/app/services/attribution_stateful_inputs.py | 276 | 9699 |
 | src/app/openapi_examples.py | 273 | 8965 |
 | src/app/services/drawdown_series.py | 271 | 8488 |
@@ -46,6 +45,7 @@ completion claim.
 | src/app/services/concentration/resolvers.py | 231 | 7566 |
 | src/app/services/rolling_metric_series.py | 221 | 6803 |
 | src/app/services/benchmark_exposure_history.py | 220 | 7264 |
+| src/app/services/calculation_supportability.py | 219 | 7214 |
 
 ### Largest Functions And Classes
 
@@ -96,12 +96,12 @@ All checks passed!
 - Ruff format check: passed
 
 ```text
-338 files already formatted
+339 files already formatted
 ```
 - Type checking: passed
 
 ```text
-Success: no issues found in 323 source files
+Success: no issues found in 324 source files
 ```
 - Unit coverage snapshot: passed
 
@@ -115,10 +115,10 @@ Success: no issues found in 323 source files
 ...
 src\app\services\rolling_stateful_source_responses.py         45      1     16      1    97%   96
 ------------------------------------------------------------------------------------------------------
-TOTAL                                                       6103     91   1028     73    98%
+TOTAL                                                       6108     91   1028     73    98%
 
-177 files skipped due to complete coverage.
-551 passed in 9.22s
+178 files skipped due to complete coverage.
+551 passed in 10.99s
 ```
 
 ## Complexity And Maintainability Snapshot
@@ -148,7 +148,7 @@ src\app\services\risk\__init__.py - A (100.00)
 - Dependency hygiene: passed
 
 ```text
-Scanning 220 files...
+Scanning 221 files...
 
 Success! No dependency issues found.
 ```
