+++ src/app/services/attribution_decomposition.py
@@ -5,7 +5,6 @@ from dataclasses import dataclass
 import pandas as pd
 
 from app.contracts.attribution import (
-    AttributionContributor,
     AttributionMetric,
     AttributionOptions,
     AttributionSetResult,
@@ -14,9 +13,11 @@ from app.contracts.attribution import (
 )
 from app.services.attribution_calculation import (
     AttributionCalculationInputs,
-    DecompositionRow,
     attribution_calculation_inputs,
-    component_decomposition,
+)
+from app.services.attribution_set_results import (
+    calculated_attribution_set,
+    empty_attribution_set,
 )
 from app.services.attribution_source_frames import (
     AttributionSourceFrames,
@@ -54,25 +55,6 @@ class AttributionSetBuildRequest:
     quality_flags: list[str]
 
 
-def _empty_attribution_set(
-    *,
-    attribution_type: AttributionType,
-    metric: AttributionMetric,
-    grouping_dimension: GroupingDimension,
-    quality_flags: list[str],
-) -> AttributionSetResult:
-    return AttributionSetResult(
-        attribution_type=attribution_type,
-        metric=metric,
-        grouping_dimension=grouping_dimension,
-        total_value=None,
-        reconciled_sum=None,
-        residual=None,
-        contributors=[],
-        quality_flags=quality_flags,
-    )
-
-
 def _unsupported_metric_set(
     *,
     attribution_type: AttributionType,
@@ -81,87 +63,20 @@ def _unsupported_metric_set(
     base_flags: list[str],
 ) -> AttributionSetResult | None:
     if attribution_type == "TOTAL_RISK" and metric != "VOLATILITY":
-        return _empty_attribution_set(
+        return empty_attribution_set(
             attribution_type=attribution_type,
             metric=metric,
             grouping_dimension=grouping_dimension,
             quality_flags=base_flags + [f"metric:{metric}:unsupported_for_total_risk"],
         )
     if attribution_type == "ACTIVE_RISK" and metric != "TRACKING_ERROR":
-        return _empty_attribution_set(
+        return empty_attribution_set(
             attribution_type=attribution_type,
             metric=metric,
             grouping_dimension=grouping_dimension,
             quality_flags=base_flags + [f"metric:{metric}:unsupported_for_active_risk"],
         )
     return None
-
-
-def _attribution_contributors(
-    *,
-    rows: list[DecompositionRow],
-    group_labels: dict[str, str | None],
-) -> list[AttributionContributor]:
-    return [
-        AttributionContributor(
-            group_key=row["group_key"],
-            group_label=group_labels.get(row["group_key"]),
-            weight_average=row["weight_average"],
-            marginal_contribution=row["marginal_contribution"],
-            component_contribution=row["component_contribution"],
-            percent_contribution=row["percent_contribution"],
-        )
-        for row in rows
-    ]
-
-
-def _reconciled_attribution_set(
-    *,
-    attribution_type: AttributionType,
-    metric: AttributionMetric,
-    grouping_dimension: GroupingDimension,
-    risk_total: float,
-    contributors: list[AttributionContributor],
-    quality_flags: list[str],
-) -> AttributionSetResult:
-    reconciled_sum = float(sum(c.component_contribution or 0.0 for c in contributors))
-    residual = float(risk_total - reconciled_sum)
-    return AttributionSetResult(
-        attribution_type=attribution_type,
-        metric=metric,
-        grouping_dimension=grouping_dimension,
-        total_value=risk_total,
-        reconciled_sum=reconciled_sum,
-        residual=residual,
-        contributors=contributors,
-        quality_flags=quality_flags,
-    )
-
-
-def _calculated_attribution_set(
-    *,
-    attribution_type: AttributionType,
-    metric: AttributionMetric,
-    grouping_dimension: GroupingDimension,
-    calculation_inputs: AttributionCalculationInputs,
-    group_labels: dict[str, str | None],
-    annualization_basis: int,
-    quality_flags: list[str],
-) -> AttributionSetResult:
-    rows = component_decomposition(
-        group_matrix=calculation_inputs.group_matrix,
-        metric_series=calculation_inputs.metric_series,
-        contribution_denominator=calculation_inputs.risk_total,
-        annualization_basis=annualization_basis,
-    )
-    return _reconciled_attribution_set(
-        attribution_type=attribution_type,
-        metric=metric,
-        grouping_dimension=grouping_dimension,
-        risk_total=calculation_inputs.risk_total,
-        contributors=_attribution_contributors(rows=rows, group_labels=group_labels),
-        quality_flags=quality_flags,
-    )
 
 
 def _attribution_set_precalculation_result(
@@ -175,7 +90,7 @@ def _attribution_set_precalculation_result(
     if calculation_inputs is None:
         return AttributionPrecalculation(
             calculation_inputs=None,
-            early_result=_empty_attribution_set(
+            early_result=empty_attribution_set(
                 attribution_type=attribution_type,
                 metric=metric,
                 grouping_dimension=grouping_dimension,
@@ -186,7 +101,7 @@ def _attribution_set_precalculation_result(
     if len(calculation_inputs.metric_series.dropna()) < 2:
         return AttributionPrecalculation(
             calculation_inputs=None,
-            early_result=_empty_attribution_set(
+            early_result=empty_attribution_set(
                 attribution_type=attribution_type,
                 metric=metric,
                 grouping_dimension=grouping_dimension,
@@ -241,7 +156,7 @@ def build_attribution_set(request: AttributionSetBuildRequest) -> AttributionSet
     if precalculation.calculation_inputs is None:
         raise RuntimeError("attribution precalculation returned no calculation inputs")
 
-    return _calculated_attribution_set(
+    return calculated_attribution_set(
         attribution_type=request.attribution_type,
         metric=request.metric,
         grouping_dimension=request.grouping_dimension,
